Reduce use of fixtures.MockPatch
I have mixed feelings about the standard library's decorators here: in
particular, the ordering requirements can be confusing, as compared to
the `useFixture` approach where you can just assign mocks directly to
variables and so it's much harder to make mistakes.  Applying a patch
for all test cases in a suite is also significantly more annoying with
the standard library's class decorator, since you have to add an
argument to every test case.  However, I've switched to using the
standard library's decorators where it's straightforward to do so.