Add a "run" command
This is still pretty basic, but it's enough to be able to launch
containers and run simple test jobs inside them.

I had to refactor the way emitter code is tested (based very loosely on
similar test code in charmcraft), since it was difficult to handle the
`emit.open_stream` context manager otherwise.