Add job matrix support
I'm not completely happy with the way this is done at the `pydantic`
level, but it works well enough for the moment.

I realized that my handling of the case where a job isn't defined for
the host architecture as an error doesn't make a lot of sense,
especially in the context of larger job matrices, so I revised that to
skip such jobs instead.