Stop running lpcraft inside the container
I borrowed the previous approach from charmcraft, and initially thought
it would be useful enough to pay its way: if we needed to wrap the job
in any complex logic that would be difficult to do at arm's length from
outside the container, then the ability to run Python code inside the
container would be helpful.

However, this approach also has a number of disadvantages: injecting
lpcraft as a snap is complex and relatively slow; it incurs all kinds of
complexity spread over lpcraft; and the split between `_run_job` and
`_run_pipeline` wasn't very clearly motivated and was confusing.

Since there doesn't turn out to be a pressing practical need for all
this complexity, remove it, and instead just execute job commands
directly inside the container.