Add a "run-one" command
We need a way to run only a single job from a pipeline, both because we
want an easy way to record log output for separate jobs separately, and
because we want to be able to retry individual jobs.

Matrix expansion makes this somewhat complicated, because as a result
there can be multiple expanded job definitions with the same name that
differ only in (for example) series or environment.  The simplest
approach I can think of turns out to be to just take an index as well as
the job name; this is safe since matrix expansion has a stable order,
and in Launchpad the dispatcher end is going to need to do its own
matrix expansion anyway.

I went for a separate "run-one" command rather than options to "run"
since the command-line syntax is sufficiently different that it makes
sense to validate it separately.