Constrain job, series, and architecture names
To avoid confusion elsewhere (for example with using these names on
command lines), it seems like a good idea to constrain job, series, and
architecture names to use a limited syntax that makes sense for unique
identifiers.  I based this on the syntax permitted for names in
Launchpad, although it isn't quite identical since I think people are
likely to want to use underscores here and there's no good reason to
forbid that.