Fix tox plugin not selecting the right interpreter
When executing tox with an envlist like `py38, py39, py310`, all three
envs were executed, but the system Python 3.8 interpreter was used for
all of them.

This seems to be an issue of the packaged tox 3.13.2 on Focal Fossa.

Upstream tox version 3.13.2 cannot be run in a Python 3.8 virtualenv at
all.

Python 3.8 support for tox was only added in 3.14.4.

This means we cannot use the packaged tox, but need to install it via
pip.