Support pipeline stages with multiple jobs
The specification calls for either of these kinds of pipeline stage to
be valid:

    pipeline:
      # this stage is a list, which means jobs are executed in parallel
      - [test, lint]
      # this stage will only execute if previous steps in the pipeline
      # passed
      - build-wheel

lpcraft's documentation agrees, but the implementation currently rejects
a pipeline stage that is a list of job names.  Implement a minimal
version of this.  The jobs aren't yet executed in parallel, but for
error handling purposes we act as if they were: all jobs in a stage are
run even if some of them fail, and we only proceed to the next stage if
they all succeed.