Isolate jobs from the project directory
Simply mounting the project directory in the container encourages jobs
to write temporary files to the project directory, which makes it far
too easy to leak files between jobs unintentionally or to corrupt the
project directory.  It's better to isolate jobs as cleanly as possible,
so give each job a fresh copy of the project directory instead.