fix config path regression
When a configurable configuration file was introduced, a default value
for the configuration file was added to the run subparser.

Running `lpcraft` without any arguments defaults to the run command, but
the subparser args do not get parsed in that case.

This has been fixed by adding an additional check/fallback to the
default configuration file.