Allow output.paths to reference parent directory
`dpkg-buildpackage` writes its output files to the parent directory of
the build tree.  This is awkward to use with lpcraft, because
`output.paths` isn't allowed to escape the build tree.

We have to work with the world as it is, not as we'd like it to be, and
it doesn't seem unreasonable to allow gathering files from the parent
directory (but not further up the tree) in order to make things easier
for people integrating build systems such as this.  Move the project
path inside the container one level down (to `/root/lpcraft/project`) to
support this.