Rearrange output directory structure
The previous structure of `<job_name>/<series>/<architecture>` didn't
work well with matrix jobs, where there can be multiple jobs with the
same series and architecture but differing by some other property.  It
also made it difficult to implement the `input` keyword, which needs to
be able to locate artifacts produced by previously-executed jobs in the
output directory; this would currently be impossible in Launchpad builds
because each individual job is run with a different `--output-directory`
in order to allow launchpad-buildd to extract just the artifacts
produced by that job.

Instead, mirror the structure implied by the `run-one` command, and
rearrange the output directory structure to `<job_name>/<job_index>`,
where `job_index` is the index of the individual job being run in the
matrix-expanded list of jobs with the same name.  This is a bit simpler,
and makes it possible to iterate over output from all jobs with the same
name just by iterating over the `<job_name>` directory.

I've checked that launchpad-buildd won't object to this rearranged
structure, although once it's in place we'll be able to change
launchpad-buildd to run all individual jobs with the same
`--output-directory` (because it will then have another way to identify
artifacts produced by individual jobs), thus preparing the way for
implementing the `input` keyword.