Fix `lpcraft run --clean`
Previously, when running two jobs with the same series and on the same
architecture, the lxc container was re-used.

After both jobs finished, it was attemted to delete "both" containers,
but that obviously failed after the first one was deleted.

Now, each container will be deleted after each job - at least when the
`--clean` flag was used.

LP: #1986374