Add --gpu-nvidia option
This allows us to pass through an NVIDIA GPU from the host system.  Jobs
will need to install some kind of CUDA binaries in order for this to be
useful.

`craft-providers` doesn't make it very straightforward to pass in
additional container configuration, but I managed to make it work by
adjusting the per-project profile before launching the container.

I've marked the option as experimental, because we need to get this in
place soon but we might find we need to change it later; in that case
coordinated changes to `lpcraft` and `launchpad-buildd` may be required.