Fix regression from adding support for support for non-LTS releases
When adding above mentioned support, also craft-providers 1.8.0 was
pulled in, which, additionally to adding the support for non-LTS
releases, also reworked the way remote images are added.

This was not uncovered as on a local developer machine the remotes were
already previously configured.

This fix has currently no explicit test, as the test would involve a
mock which returns a mock on which we could assert that a method was
called.