Fix default value for the `root` flag in documentation
The default value for the `root` flag was listed as `true`, but
implemented as `false`.

`true` is a saner value as there might be quite some existing
configurations out there relying on that, and we certainly do not want
to break them.

Also see the intital discussion
https://bugs.launchpad.net/lpci/+bug/1982954
and the implementation
https://code.launchpad.net/~pelpsi/lpci/+git/lpcraft/+merge/441539

LP: #2025359