Release the latest build of each architecture
`lpci release` incorrectly released the latest build regardless of
architecture; this approach was OK when builds were typically only
dispatched for a single architecture, but now that we need to handle
releasing for (e.g.) both amd64 and arm64, it doesn't work so well.
Release the latest build for each architecture for which builds exist
instead, and add a new `--architecture` option for the case where people
want to override this behaviour and only release the latest build for a
single architecture.