Add support for 23.10 (mantic) and drop 22.10 (kinetic)
This requires an update to depend on craft-providers>=1.19.0.
requirements.txt is regeneraged as a result.

The environment and hostname usage in LPCIBuilddBaseConfiguration is
updated based on craft-providers refactoring of base handling code.

Unit tests are updated with timeout argument that was added in
craft-providers refactoring of base handling code.

Also fix mypy complaint:
'error: Unused "type: ignore" comment  [unused-ignore]'