Fix the build error with snapcraft 8.x
There was a bug in the sitecustomize.py file bundled with the lpci snap,
which caused the build process to depend on the version of python used
by snapcraft. This bug didn't surface with snapcraft 7.x, which is built
on core20, the same as lpci snap. But snapcraft 8.x is built on core22
and hence there was a conflict between the expected and available versions
of pip during the build process.

This fixes the issue by enabling the site customizations only in the
context of the lpci snap.

LP: #2053109